+++ agent/src/main/java/com/teamscale/jacoco/agent/options/AgentOptionsParser.java
@@ -15,27 +15,18 @@ import com.teamscale.report.util.BashFileSkippingInputStream;
 import com.teamscale.report.util.ILogger;
 
 import org.conqat.lib.commons.collections.CollectionUtils;
-import org.conqat.lib.commons.collections.Pair;
-import org.conqat.lib.commons.filesystem.AntPatternUtils;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.InvalidPathException;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.List;
-import java.util.function.Predicate;
 import java.util.jar.JarInputStream;
 import java.util.jar.Manifest;
-import java.util.regex.Pattern;
-
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toList;
 
 import okhttp3.HttpUrl;
 
@@ -47,17 +38,12 @@ public class AgentOptionsParser {
 	/** Character which starts a comment in the config file. */
 	private static final String COMMENT_PREFIX = "#";
 
-	/** Stand-in for the question mark operator. */
-	private static final String QUESTION_REPLACEMENT = "!@";
-
-	/** Stand-in for the asterisk operator. */
-	private static final String ASTERISK_REPLACEMENT = "#@";
-
-	/** Logger. */
 	private final ILogger logger;
+	private final FilePatternResolver filePatternResolver;
 
 	public AgentOptionsParser(ILogger logger) {
 		this.logger = logger;
+		this.filePatternResolver = new FilePatternResolver(logger);
 	}
 
 	/**
@@ -133,10 +119,10 @@ public class AgentOptionsParser {
 									   String value) throws AgentOptionParseException {
 		switch (key) {
 			case "config-file":
-				readConfigFromFile(options, parseFile(key, value));
+				readConfigFromFile(options, filePatternResolver.parsePath(key, value).toFile());
 				return true;
 			case "logging-config":
-				options.loggingConfig = parsePath(key, value);
+				options.loggingConfig = filePatternResolver.parsePath(key, value);
 				return true;
 			case "interval":
 				try {
@@ -149,7 +135,7 @@ public class AgentOptionsParser {
 				options.validateSsl = Boolean.parseBoolean(value);
 				return true;
 			case "out":
-				options.setParentOutputDirectory(parsePath(key, value));
+				options.setParentOutputDirectory(filePatternResolver.parsePath(key, value));
 				return true;
 			case "upload-url":
 				options.uploadUrl = parseUrl(value);
@@ -181,8 +167,9 @@ public class AgentOptionsParser {
 				options.jacocoExcludes = value.replaceAll(";", ":");
 				return true;
 			case "class-dir":
-				options.classDirectoriesOrZips = CollectionUtils
-						.mapWithException(splitMultiOptionValue(value), singleValue -> parseFile(key, singleValue));
+				List<String> list = splitMultiOptionValue(value);
+				options.classDirectoriesOrZips = ClasspathUtils
+						.resolveClasspathTextFiles(key, filePatternResolver, list);
 				return true;
 			default:
 				return false;
@@ -244,7 +231,8 @@ public class AgentOptionsParser {
 				options.teamscaleServer.commit = parseCommit(value);
 				return true;
 			case AgentOptions.TEAMSCALE_COMMIT_MANIFEST_JAR_OPTION:
-				options.teamscaleServer.commit = getCommitFromManifest(parseFile(key, value));
+				options.teamscaleServer.commit = getCommitFromManifest(
+						filePatternResolver.parsePath(key, value).toFile());
 				return true;
 			case AgentOptions.TEAMSCALE_GIT_PROPERTIES_JAR_OPTION:
 				options.teamscaleServer.revision = parseGitProperties(key, value);
@@ -266,7 +254,7 @@ public class AgentOptionsParser {
 	}
 
 	private String parseGitProperties(String key, String value) throws AgentOptionParseException {
-		File jarFile = parseFile(key, value);
+		File jarFile = filePatternResolver.parsePath(key, value).toFile();
 		try {
 			String commit = GitPropertiesLocator.getCommitFromGitProperties(jarFile);
 			if (commit == null) {
@@ -357,108 +345,6 @@ public class AgentOptionsParser {
 		}
 	}
 
-	/**
-	 * Parses the given value as a {@link File}.
-	 */
-	private File parseFile(String optionName, String value) throws AgentOptionParseException {
-		return parsePath(optionName, new File("."), value).toFile();
-	}
-
-	/**
-	 * Parses the given value as a {@link Path}.
-	 */
-	private Path parsePath(String optionName, String value) throws AgentOptionParseException {
-		return parsePath(optionName, new File("."), value);
-	}
-
-	/**
-	 * Parses the given value as a {@link Path}.
-	 */
-	/* package */ Path parsePath(String optionName, File workingDirectory,
-								 String value) throws AgentOptionParseException {
-		if (isPathWithPattern(value)) {
-			return parseFileFromPattern(workingDirectory, optionName, value);
-		}
-		try {
-			return workingDirectory.toPath().resolve(Paths.get(value));
-		} catch (InvalidPathException e) {
-			throw new AgentOptionParseException("Invalid path given for option " + optionName + ": " + value, e);
-		}
-	}
-
-	/** Parses the value as a ant pattern to a file or directory. */
-	private Path parseFileFromPattern(File workingDirectory, String optionName,
-									  String value) throws AgentOptionParseException {
-		Pair<String, String> baseDirAndPattern = splitIntoBaseDirAndPattern(value);
-		String baseDir = baseDirAndPattern.getFirst();
-		String pattern = baseDirAndPattern.getSecond();
-
-		File workingDir = workingDirectory.getAbsoluteFile();
-		Path basePath = workingDir.toPath().resolve(baseDir).normalize().toAbsolutePath();
-
-		Pattern pathMatcher = AntPatternUtils.convertPattern(pattern, false);
-		Predicate<Path> filter = path -> pathMatcher
-				.matcher(FileSystemUtils.normalizeSeparators(basePath.relativize(path).toString())).matches();
-
-		List<Path> matchingPaths;
-		try {
-			matchingPaths = Files.walk(basePath).filter(filter).sorted().collect(toList());
-		} catch (IOException e) {
-			throw new AgentOptionParseException(
-					"Could not recursively list files in directory " + basePath + " in order to resolve pattern " + pattern + " given for option " + optionName,
-					e);
-		}
-
-		if (matchingPaths.isEmpty()) {
-			throw new AgentOptionParseException(
-					"Invalid path given for option " + optionName + ": " + value + ". The pattern " + pattern +
-							" did not match any files in " + basePath.toAbsolutePath());
-		} else if (matchingPaths.size() > 1) {
-			logger.warn(
-					"Multiple files match the pattern " + pattern + " in " + basePath
-							.toString() + " for option " + optionName + "! " +
-							"The first one is used, but consider to adjust the " +
-							"pattern to match only one file. Candidates are: " + matchingPaths.stream()
-							.map(basePath::relativize).map(Path::toString).collect(joining(", ")));
-		}
-		Path path = matchingPaths.get(0).normalize();
-		logger.info("Using file " + path + " for option " + optionName);
-
-		return path;
-	}
-
-	/**
-	 * Splits the path into a base dir, a the directory-prefix of the path that does not contain any ? or *
-	 * placeholders, and a pattern suffix. We need to replace the pattern characters with stand-ins, because ? and * are
-	 * not allowed as path characters on windows.
-	 */
-	private Pair<String, String> splitIntoBaseDirAndPattern(String value) {
-		String pathWithArtificialPattern = value.replace("?", QUESTION_REPLACEMENT).replace("*", ASTERISK_REPLACEMENT);
-		Path pathWithPattern = Paths.get(pathWithArtificialPattern);
-		Path baseDir = pathWithPattern;
-		while (isPathWithArtificialPattern(baseDir.toString())) {
-			baseDir = baseDir.getParent();
-			if (baseDir == null) {
-				return new Pair<>("", value);
-			}
-		}
-		String pattern = baseDir.relativize(pathWithPattern).toString().replace(QUESTION_REPLACEMENT, "?")
-				.replace(ASTERISK_REPLACEMENT, "*");
-		return new Pair<>(baseDir.toString(), pattern);
-	}
-
-	/** Returns whether the given path contains ant pattern characters (?,*). */
-	private static boolean isPathWithPattern(String path) {
-		return path.contains("?") || path.contains("*");
-	}
-
-	/**
-	 * Returns whether the given path contains artificial pattern characters ({@link #QUESTION_REPLACEMENT}, {@link
-	 * #ASTERISK_REPLACEMENT}).
-	 */
-	private static boolean isPathWithArtificialPattern(String path) {
-		return path.contains(QUESTION_REPLACEMENT) || path.contains(ASTERISK_REPLACEMENT);
-	}
 
 	/**
 	 * Parses the given value as a URL or returns <code>null</code> if that fails.